Refactor AppState into appstate.rs
