+++ src/lib.rs
@@ -1,3 +1,4 @@
+pub mod appstate;
 pub mod auth;
 pub mod dashboard_server;
 pub mod dispatcher;
@@ -5,12 +6,10 @@ pub mod health;
 pub mod protocol;
 pub mod utils;
 
+pub use appstate::{AppState, LogBuffer, LogBufferWriter};
 pub use auth::UserRegistry;
 pub use dashboard_server::DashboardServer;
-pub use dispatcher::{
-    AppState, LogBuffer, LogBufferWriter, model_handler, models_handler, proxy_handler, run_worker,
-    tags_handler,
-};
+pub use dispatcher::{model_handler, models_handler, proxy_handler, run_worker, tags_handler};
 pub use health::health_handler;
 pub use protocol::{
     BackendSnapshot, DashboardCmd, DashboardSnapshot, consumed_len, decode, encode,
